// src/twitchAPI-types.ts
export interface DBTwitchUser {
  id: string;
  login: string;
  display_name: string;
  type: string;
  broadcaster_type: string;
  description: string;
  profile_image_url: string;
  offline_image_url: string;
  view_count: number;
  email: string;
  created_at: string;
  banned_date?: string;
  permitted_date?: string;
}
// ...
export interface MessageData {
  user: {
    id: string;
    name: string;
    login: string;
    colorRaw: string;
    badges: TwitchBadgesList;
    badgesRaw: string;
    color?: Color;
    perms: {
      mod: boolean;
      broadcaster: boolean;
      vip: boolean;
      owner: boolean;
    };
  };
  channel: {
    id: string;
    login: string;
  };
  messageID: string;
  isAction: boolean;
  raw: string;
  text: string;
  timestamp: string;
  emotes: TwitchEmoteList;
  tags: IRCMessageTags;
  bits?: number;
  args?: string[];
  prefix: string;
  commandName?: string;
  send: (message: string) => Promise<void>;
}
// ...
export interface TwitchViewers {
  broadcaster: string[];
  vips: string[];
  moderators: string[];
  staff: string[];
  admins: string[];
  global_mods: string[];
  viewers: string[];
}

export interface StreamData {
  raiders: string[];
  subs: string[];
  reSubs: string[];
  gifters: any;
  // seedBank: liveData.seedBank ?? [],
  welcomeList: string[];
  cheerers: string[];
  binxRaids: string[];
}